change mod struct
